pyproject.toml
[project]
name = "stock"
version = "0.1.0"
description = "Investment Intelligence Analyst"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "python-dotenv>=1.2.2",
    "yfinance>=1.3.0",
]

[project.optional-dependencies]
local = [
    "anthropic>=0.97.0",
    "mlx-whisper>=0.4.3",
    "ollama>=0.6.2",
    "playwright>=1.59.0",
    "sentence-transformers>=5.4.1",
]
